c90: Update lexer

+++ pkg/c90/y.go
@@ -189,7 +189,7 @@ const yyErrCode = 2
 const yyInitialStackSize = 16
 
 //line yacctab:1
-var yyExca = [...]int8{
+var yyExca = [...]int{
 	-1, 1,
 	1, -1,
 	-2, 0,
@@ -199,7 +199,7 @@ const yyPrivate = 57344
 
 const yyLast = 1280
 
-var yyAct = [...]int16{
+var yyAct = [...]int{
 	72, 95, 94, 296, 145, 6, 233, 93, 121, 75,
 	149, 41, 47, 5, 5, 226, 133, 123, 116, 11,
 	48, 49, 50, 115, 201, 140, 10, 144, 139, 308,
@@ -330,7 +330,7 @@ var yyAct = [...]int16{
 	24, 25, 22, 23, 29, 30, 17, 36, 37, 35,
 }
 
-var yyPact = [...]int16{
+var yyPact = [...]int{
 	926, 926, -1000, -1000, -1000, 28, 498, 1231, 1231, 1231,
 	310, 236, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
@@ -368,7 +368,7 @@ var yyPact = [...]int16{
 	-1000, -1000, 580, -41, -1000, 580, -1000, -1000, -1000,
 }
 
-var yyPgo = [...]int16{
+var yyPgo = [...]int{
 	0, 428, 427, 34, 426, 419, 7, 1, 418, 8,
 	24, 35, 44, 51, 45, 18, 23, 64, 67, 80,
 	417, 2, 416, 17, 301, 12, 415, 405, 50, 236,
@@ -378,7 +378,7 @@ var yyPgo = [...]int16{
 	106, 505, 353, 354,
 }
 
-var yyR1 = [...]int8{
+var yyR1 = [...]int{
 	0, 2, 2, 2, 2, 4, 4, 4, 4, 4,
 	4, 4, 4, 5, 5, 7, 7, 7, 7, 7,
 	7, 8, 8, 8, 8, 8, 8, 9, 9, 11,
@@ -403,7 +403,7 @@ var yyR1 = [...]int8{
 	63, 63,
 }
 
-var yyR2 = [...]int8{
+var yyR2 = [...]int{
 	0, 1, 1, 1, 3, 1, 4, 3, 4, 3,
 	3, 2, 2, 1, 3, 1, 2, 2, 2, 2,
 	4, 1, 1, 1, 1, 1, 1, 1, 4, 1,
@@ -428,7 +428,7 @@ var yyR2 = [...]int8{
 	3, 2,
 }
 
-var yyChk = [...]int16{
+var yyChk = [...]int{
 	-1000, -1, -62, -63, -24, -25, -31, -27, -28, -29,
 	-43, -44, 30, 31, 32, 33, 34, 45, 35, 36,
 	37, 38, 41, 42, 39, 40, -33, -34, 29, 43,
@@ -466,7 +466,7 @@ var yyChk = [...]int16{
 	65, 63, 53, 63, -53, 63, -53, 83, -53,
 }
 
-var yyDef = [...]int16{
+var yyDef = [...]int{
 	0, -2, 204, 206, 207, 0, 0, 77, 79, 81,
 	0, 131, 87, 88, 89, 90, 91, 92, 93, 94,
 	95, 96, 97, 98, 99, 100, 101, 102, 103, 128,
@@ -504,7 +504,7 @@ var yyDef = [...]int16{
 	163, 167, 0, 0, 197, 0, 193, 196, 198,
 }
 
-var yyTok1 = [...]int8{
+var yyTok1 = [...]int{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -520,7 +520,7 @@ var yyTok1 = [...]int8{
 	3, 3, 3, 84, 79, 85, 72,
 }
 
-var yyTok2 = [...]int8{
+var yyTok2 = [...]int{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -529,7 +529,7 @@ var yyTok2 = [...]int8{
 	52, 53, 54, 55, 56, 57, 58, 59, 60, 61,
 }
 
-var yyTok3 = [...]int8{
+var yyTok3 = [...]int{
 	0,
 }
 
@@ -611,9 +611,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := int(yyPact[state])
+	base := yyPact[state]
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
+		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -623,13 +623,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
+		for yyExca[i] != -1 || yyExca[i+1] != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := int(yyExca[i])
+			tok := yyExca[i]
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -660,30 +660,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = int(yyTok1[0])
+		token = yyTok1[0]
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = int(yyTok1[char])
+		token = yyTok1[char]
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = int(yyTok2[char-yyPrivate])
+			token = yyTok2[char-yyPrivate]
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = int(yyTok3[i+0])
+		token = yyTok3[i+0]
 		if token == char {
-			token = int(yyTok3[i+1])
+			token = yyTok3[i+1]
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = int(yyTok2[1]) /* unknown char */
+		token = yyTok2[1] /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -738,7 +738,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = int(yyPact[yystate])
+	yyn = yyPact[yystate]
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -749,8 +749,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = int(yyAct[yyn])
-	if int(yyChk[yyn]) == yytoken { /* valid shift */
+	yyn = yyAct[yyn]
+	if yyChk[yyn] == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -763,7 +763,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = int(yyDef[yystate])
+	yyn = yyDef[yystate]
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -772,18 +772,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
+			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = int(yyExca[xi+0])
+			yyn = yyExca[xi+0]
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = int(yyExca[xi+1])
+		yyn = yyExca[xi+1]
 		if yyn < 0 {
 			goto ret0
 		}
@@ -805,10 +805,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
+				yyn = yyPact[yyS[yyp].yys] + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
-					if int(yyChk[yystate]) == yyErrCode {
+					yystate = yyAct[yyn] /* simulate a shift of "error" */
+					if yyChk[yystate] == yyErrCode {
 						goto yystack
 					}
 				}
@@ -844,7 +844,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= int(yyR2[yyn])
+	yyp -= yyR2[yyn]
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -855,16 +855,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = int(yyR1[yyn])
-	yyg := int(yyPgo[yyn])
+	yyn = yyR1[yyn]
+	yyg := yyPgo[yyn]
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = int(yyAct[yyg])
+		yystate = yyAct[yyg]
 	} else {
-		yystate = int(yyAct[yyj])
-		if int(yyChk[yystate]) != -yyn {
-			yystate = int(yyAct[yyg])
+		yystate = yyAct[yyj]
+		if yyChk[yystate] != -yyn {
+			yystate = yyAct[yyg]
 		}
 	}
 	// dummy call; replaced with literal code
